refactor: AddDataSourceModalのロジックをComposableに移植、URLバリデーションをリアルタイム化
- AddDataSourceModal.vueのscriptタグ内のロジックをuse-add-data-source-modal.tsに移植
- URLバリデーションをURL入力と同時に行うように変更（validateUrl関数をwatchで呼び出し）
- AddDataSourceModal.vueを簡素化し、composableを使用するように修正
- .editorconfigを追加
